Build path to indicator data file

Builds path to indicator files of the form /{base_dir}/data-raw/indicator/{keyword}_{suffix}.csv.

path_keyword(keyword, geo = "CH", suffix = "sa")

Arguments

keyword

Keyword (character vector) for which to construct the path to the indicator.

geo

geographical area

suffix

Character vector for file suffix.

See also