Download daily, weekly and monthly Google Trends data for a keyword.

Downloads daily, weekly and monthly Google Trends data for a keyword and writes the data to csv files.

proc_keyword_init(
  keyword = "Insolvenz",
  geo = "CH",
  from = "2006-01-01",
  indicator_raw = FALSE
)

Arguments

keyword

A single keyword to query Google Trends.

geo

A character vector denoting the geographic region. Default is "CH".

from

Start of timeframe in YYYY-mm-dd form. Should not be changed from the the default.

indicator_raw

store individual downloads. If FALSE, only the averages are stored.

Details

By default, the data is stored in folders data and raw. Each folder contains a subdirectory of each country.

See also

Examples

if (FALSE) {
# run once
proc_keyword_init(keyword = "Insolvenz", "AT")
# run every day
proc_keyword_latest(keyword = "Insolvenz", "AT")
}